% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_model.R
\name{compile_model}
\alias{compile_model}
\title{Compile a brms model}
\usage{
compile_model(model_data, backend = "cmdstanr")
}
\arguments{
\item{model_data}{\code{tibble::tibble} result of running \code{prepare_model_data}}
}
\value{
\code{brms::brmsfit} object
}
\description{
Given a dataset that has been prepared for model to predict the number of
positive, cells given a matrix of cell counts for each gene in each well,
but don't run it.
}
\examples{
\dontrun{
.   data <- simulate_screen(...)
.   model_data <- prepare_model_data(data)
.   model <- compile_model(model_data)
.   model_fit <- fit_model(model_data, model)
.   model_estimate <- gather_model_estimate(model_fit)
.   model_evaluation <- evaluate_model_fit(data, model_estimate)
}

}
