% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_model.R
\name{evaluate_model_fit}
\alias{evaluate_model_fit}
\title{Evaluate the model performance against ground truth hit/no-hit status}
\usage{
evaluate_model_fit(data, model_estimate)
}
\arguments{
\item{data}{\code{data.frame} e.g. generated by \code{simulate_data} having columns
[\code{gene}, hit`]}

\item{model_estimate}{\code{data.frame} e.g. produced by \code{gather_model_estimate}
having columns [\code{gene}, \code{mean}]}
}
\value{
\code{data.frame} with columns [\code{model_ap}, \code{model_auroc}]
}
\description{
Evaluate the average precision and area under the ROC curve as quality
metrics for predicting the hit-status of each gene.
}
\examples{
\dontrun{
  data <- simulate_screen(...)
  model_data <- prepare_model_data(data)
  model <- compile_model(model_data)
  model_fit <- fit_model(model_data, model)
  model_estimate <- gather_model_estimate(model_fit)
  model_evaluation <- evaluate_model_fit(data, model_estimate)
}


}
