% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_model.R
\name{fit_model}
\alias{fit_model}
\title{Fit a brms model}
\usage{
fit_model(
  model_data,
  model,
  chains = 4,
  iter = 8000,
  cores = 4,
  control = list(adapt_delta = 0.99, max_treedepth = 12),
  ...
)
}
\arguments{
\item{model_data}{\code{tibble::tibble} data prepared for the model with
\code{model_prepare_data}}

\item{model}{\code{brms::brmfit} a compiled model the result of \code{compile_model}}

\item{chains}{int number of parellel chains to run}

\item{iter}{int number of iterations for each chain}

\item{cores}{int number of cores to use (e.g. the same number of chains)}

\item{control}{list parameters to control the mcmc sampling}
}
\value{
\code{brms::brmsfit} object
}
\description{
Given a model prepared by \code{compile_model} fit it with the given data
}
\details{
Note by separating out the compilation step it is one model can be
fit multiple times.
}
