% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_classification_scores.R
\name{plot_classification_scores}
\alias{plot_classification_scores}
\title{Plot Classification Scores}
\usage{
plot <- plot_classification_scores(
     class_pos_mu = 0.99,
     class_pos_var = 0.0001,
     class_neg_mu = 0.01,
     class_neg_var = 0.0001)
   ggplot2::ggsave(
     filename = "classification_scores.png",
     plot = plot,
     width = 5,
     height = 4)
}
\arguments{
\item{class_pos_mu}{numeric mean classification score of the negative class}

\item{class_pos_var}{numeric variance of classification score of the negative
class}
}
\value{
\code{ggplot2::ggplot} object.
}
\description{
Given parameters for the distribution of scores for negative and positively
scored objects, generate a plot of the distribution of scores for each class
}
\details{
The functional for each class is a beta distribution, where the parameters
are given in the mean and variance form.
}
