% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_genes_per_well.R
\name{plot_genes_per_well}
\alias{plot_genes_per_well}
\title{Plot genes per well histogram}
\usage{
plot_genes_per_well(data)
}
\arguments{
\item{data}{data.frame with columns [\code{well}, \code{gene}] and each row is a well
well x gene pair}
}
\value{
\code{ggplot2::ggplot} object
}
\description{
From the given dataset, plot the distribution of genes per well
as a histogram
}
\examples{
\dontrun{
   data <- simulate_data(...)
   plot <- plot_genes_per_well(data)
   ggplot2::ggsave(
     filename = "genes_per_well.png",
     plot = plot,
     width = 5,
     height = 4)
}

}
