% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_gene_estimates.R
\name{plot_model_estimate}
\alias{plot_model_estimate}
\title{Plot model estimate}
\usage{
plot_model_estimate(model_estimate)
}
\arguments{
\item{model_estimate}{data.frame result of calling \code{gather_model_estimate}}
}
\value{
\code{ggplot2::ggplot} object
}
\description{
For each gene (vertical) plot the model estimate (horizontal) as
the mean and 95\% credible interval
}
\examples{
\dontrun{
   data <- simulate_data(...)
   model_data <- prepare_model_data(data)
   model <- compile_model(model_data)
   model_fit <- fit_model(model_data, model)
   model_estimate <- gather_model_estimate(model_fit)
   plot <- plot_model_estimate(model_estimate)
   ggplot2::ggsave(
     filename = "model_estimate.png",
     plot = plot,
     width = 5,
     height = 4)
}

}
