% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_parameter_scan.R
\name{plot_parameter_scan}
\alias{plot_parameter_scan}
\title{Plot results of a parameter scan}
\usage{
plot_parameter_scan(
  parameter_scan,
  param_x = "n_wells_per_screen",
  param_color = "n_genes_per_well",
  param_y = "model_auroc",
  label_x = "N Wells Per Screen",
  label_y = "Model AUROC (larger is better)",
  label_color = "N Genes Per Well",
  title = "Parameter Scan"
)
}
\arguments{
\item{parameter_scan}{data.frame with columns specified by the parameters
[\verb{<param_x>}, \verb{<param_y>}, \verb{<param_color>}].}

\item{param_x}{character column name of parameter to be mapped to the x-axis}

\item{param_color}{character column name of parameter to be mapped to the
color dimension}

\item{param_y}{character column name of parameter to be mapped to the y-axis}
}
\value{
\code{ggplot2::ggplot} object
}
\description{
Given a dataframe with parameters and scores, plot the scores
across specified parameters
}
\examples{
\dontrun{
  parameter_scan <- readr::read_tsv(file = "parameter_scan.tsv")
  plot <- plot_parameter_scan(parameter_scan)
  ggplot2::ggsave(
    filename = "parameter_scan_plot.png",
    width = 6,
    height = 5)
}


}
