% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_positivity_rate_by_well.R
\name{plot_positivity_rate_by_well}
\alias{plot_positivity_rate_by_well}
\title{Plot positivity rate by well}
\usage{
plot_positivity_rate_by_well(data)
}
\arguments{
\item{data}{data.frame with columns [\code{well}, \code{n_cells_per_gene_per_well},
\code{positive}]}
}
\value{
\code{ggplot2::ggplot} object
}
\description{
Given a dataset generate a scatter plot comparing number of positive
cells by the number of total cells for each well.
}
\examples{
\dontrun{
   data <- simulate_data(...)
   plot <- plot_positivity_rate_by_well(data)
   ggplot2::ggsave(
     filename = "positivity_rate_by_well.png",
     plot = plot,
     width = 5,
     height = 4)
}

}
