% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_well_gene_reads_heatmap.R
\name{plot_well_gene_reads_heatmap}
\alias{plot_well_gene_reads_heatmap}
\title{Plot well x gene Reads heatmap}
\usage{
plot_well_gene_reads_heatmap(data)
}
\arguments{
\item{data}{data.frame with columns [\code{gene}, \code{well},
\code{n_reads_per_gene_per_well}]}
}
\value{
\code{ggplot2::ggplot} object
}
\description{
Given a dataset plot heatmap where each box corresponds to a well
(row) by gene (column) pair. The values are the \code{log10(reads + 1)} on
a scale from blue to yellow.
}
\examples{
\dontrun{
   data <- simulate_screen(...)
   plot <- plot_well_gene_reads_heatmap(data)
   ggplot2::ggsave(
     filename = "well_gene_reads_heatmap.png",
     plot = plot,
     width = 5,
     height = 4)
}

}
