% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_wells_per_gene.R
\name{plot_wells_per_gene}
\alias{plot_wells_per_gene}
\title{Plot wells per gene histogram}
\usage{
plot_wells_per_gene(data)
}
\arguments{
\item{data}{data.frame with columns [\code{well}, \code{gene}] and each row is a well
well x gene pair}
}
\value{
\code{ggplot2::ggplot} object
}
\description{
From the given dataset, plot the distribution of wells per gene
as a histogram
}
\examples{
\dontrun{
   data <- simulate_data(...)
   plot <- plot_wells_per_gene(data)
   ggplot2::ggsave(
     filename = "wells_per_gene.png",
     plot = plot,
     width = 5,
     height = 4)
}

}
