% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_model.R
\name{prepare_model_data}
\alias{prepare_model_data}
\title{Prepare data for modeling}
\usage{
prepare_model_data(data)
}
\arguments{
\item{data}{data.frame with columns [\code{well}, \code{gene}, \code{positive},
\code{n_cells_per_gene_per_well}]}
}
\value{
\code{tibble::tibble} with columns [\code{well}, \code{Npositive}, \code{log10counts}]
where the types are \code{factor}, \code{numeric}, \code{matrix} and the matrix has
\code{1} row and \code{n_gene} columns.
}
\description{
Given a dataset, compute for each (well, gene) pair, the number of positive
and log10 total cell count.
}
\examples{
\dontrun{
  data <- simulate_screen(...)
  model_data <- prepare_model_data(data)
  model <- compile_model(model_data)
  model_fit <- fit_model(model_data, model)
  model_estimate <- gather_model_estimate(model_fit)
  model_evaluation <- evaluate_model_fit(data, model_estimate)
}



}
