% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scan_parameters.R
\name{scan_parameters}
\alias{scan_parameters}
\title{Scan and evaluate simulated screens}
\usage{
scan_parameters(
  n_genes_per_library,
  gene_abundance_factor_mu,
  gene_abundance_factor_var,
  gene_hit_rate,
  n_wells_per_screen,
  well_abundance_factor_mu,
  well_abundance_factor_var,
  n_genes_per_well,
  n_cells_per_well_mu,
  n_cells_per_well_var,
  class_pos_mu,
  class_pos_var,
  class_neg_mu,
  class_neg_var,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{n_genes_per_library}{integer number of genes in the library}

\item{gene_abundance_factor_mu}{numeric}

\item{gene_abundance_factor_var}{numeric}

\item{gene_hit_rate}{numeric}

\item{n_wells_per_screen}{integer}

\item{well_abundance_factor_mu}{numeric}

\item{well_abundance_factor_var}{numeric}

\item{n_genes_per_well}{integer}

\item{n_cells_per_well_mu}{numeric}

\item{n_cells_per_well_var}{numeric}

\item{class_pos_mu}{numeric}

\item{class_pos_var}{numeric}

\item{class_neg_mu}{numeric}

\item{class_neg_var}{numeric}

\item{verbose}{= FALSE}
}
\description{
For each combination of given parameters, generate a simulated dataset
fit a model to it and estimate the predictive accuracy.
}
