% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scan_parameters.R
\name{scan_parameters_future}
\alias{scan_parameters_future}
\title{Scan and evaluate simulated screens using future package}
\usage{
scan_parameters_future(
  n_genes_in_library,
  gene_abundance_alpha,
  gene_hit_rate,
  n_wells_per_screen,
  well_abundance_factor_mu,
  well_abundance_factor_var,
  imaging_n_cells_per_well_lambda,
  imaging_n_cells_per_well_nu,
  class_pos_mu,
  class_pos_var,
  class_neg_mu,
  class_neg_var,
  sequencing_n_cells_per_well_lambda,
  sequencing_n_cells_per_well_nu,
  pcr_factor_mu,
  pcr_factor_var,
  n_reads_total,
  progress_file = NULL,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{n_genes_in_library}{integer number of genes in the library}

\item{gene_abundance_alpha}{numeric}

\item{gene_hit_rate}{numeric}

\item{n_wells_per_screen}{integer}

\item{well_abundance_factor_mu}{numeric}

\item{well_abundance_factor_var}{numeric}

\item{imaging_n_cells_per_well_lambda}{numeric,}

\item{imaging_n_cells_per_well_nu}{numeric,}

\item{class_pos_mu}{numeric}

\item{class_pos_var}{numeric}

\item{class_neg_mu}{numeric}

\item{class_neg_var}{numeric}

\item{sequencing_n_cells_per_well_lambda}{numeric}

\item{sequencing_n_cells_per_well_nu}{numeric}

\item{pcr_factor_mu}{numeric}

\item{pcr_factor_var}{numeric}

\item{n_reads_total}{numeric}

\item{progress_file}{numeric}

\item{verbose}{bool}
}
\description{
For each combination of given parameters, generate a simulated dataset
fit a model to it and estimate the predictive accuracy.
}
\examples{
\dontrun{

scores <- spaCRPower::scan_parameters(
    n_genes_in_library = 200,
    gene_abundance_alpha = 1000,
    gene_hit_rate = 0.05,
    n_wells_per_screen = 20,
    well_abundance_factor_mu = 10,
    well_abundance_factor_var = 2,
    imaging_n_cells_per_well_lambda = 1000,
    imaging_n_cells_per_well_nu = 1,
    class_pos_mu = 0.99,
    class_pos_var = .0001,
    class_neg_mu = 0.01,
    class_neg_var = .0001,
    sequencing_n_cells_per_well_lambda = 1000,
    sequencing_n_cells_per_well_nu = 1,
    pcr_factor_mu = 0.1,
    pcr_factor_var = 0.01,
    n_reads_total = 1000,
    progress_file = "parameter_scan.tsv",
    verbose = TRUE,
    refresh = 0)
}

}
