% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_screen.R
\name{simulate_imaging_plate}
\alias{simulate_imaging_plate}
\title{Simulate an imaging plate}
\usage{
simulate_imaging_plate(
  spot_plate,
  imaging_n_cells_per_well_lambda,
  imaging_n_cells_per_well_nu,
  class_pos_mu,
  class_pos_var,
  class_neg_mu,
  class_neg_var
)
}
\arguments{
\item{class_pos_mu}{numeric mean of the probability of being called positive
for cells that are hits}

\item{class_pos_var}{numeric variance of the probability of being called
positive for cells that are hits}

\item{class_neg_mu}{numeric mean of the probability of being called positive
for cells that are not hits}

\item{class_neg_var}{numeric variance of the probability of being called
positive for cells that are not hits}

\item{spotplate}{data.frame results of calling \code{simulate_spot_plate()}}

\item{n_cells_per_well_lambda}{numeric mean of the number of cells for each
gene in each well (must be greater than 0)}

\item{n_cells_per_well_nu}{numeric dispersion of the number of wells for each
per-well number of cells}
}
\value{
data.frame
}
\description{
Simulate an imaging plate
}
