% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_screen.R
\name{simulate_spot_plate}
\alias{simulate_spot_plate}
\title{Simulate a spot plate}
\usage{
simulate_spot_plate(
  gene_library,
  n_wells_per_screen,
  well_abundance_factor_mu,
  well_abundance_factor_var
)
}
\arguments{
\item{gene_library}{data.frame result of calling \code{simulate_library()}}

\item{n_wells_per_screen}{integer number of wells in the screen}

\item{well_abundance_factor_var}{numeric variance of the per-well abundance
factor (must be greater than 0).}

\item{well_abundance_facotr_mu}{numeric mean of the per-well abundance
factor (must be greater than 0).}
}
\value{
data.frame with one row for each gene, well pair and columns
\verb{\[gene, well, <gene_covariates>, <well_covariates>, gene_in_well\]}, where
the \code{gene} and \code{well} columns are 1-based indices for each gene and well,
\verb{<gene_covariates>} are columns from the inpute \code{gene_library},
\verb{<well_covariates>} are the \code{well_abundance_factor_mu} parameters
\code{well_abundance_factor_var}, and the sampled \code{well_abundance} value.
}
\description{
Given the number of wells and parameters for the well abundance, simulate
whether or not gene i is in well j for all i and j.
}
\details{
well_abundance_j ~ gamma(
mean = well_abundance_factor_mu,
var = well_abundance_factor_var)
gene_in_well_ij ~ binom(gene_abundance_i * well_abundance_j)

Note we use \code{rgamma_mean_variance()} to get the gamma distribution
parameterized by the mean and variance.
}
