% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scan_parameters.R
\name{scan_parameters}
\alias{scan_parameters}
\title{Scan and evaluate simulated screens}
\usage{
scan_parameters(
  n_genes_per_library,
  gene_abundance_factor_mu,
  gene_abundance_factor_var,
  gene_hit_rate,
  n_wells_per_screen,
  well_abundance_factor_mu,
  well_abundance_factor_var,
  n_genes_per_well,
  n_cells_per_well_mu,
  n_cells_per_well_var,
  n_reads_total,
  class_pos_mu,
  class_pos_var,
  class_neg_mu,
  class_neg_var,
  progress_file = NULL,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{n_genes_per_library}{integer number of genes in the library}

\item{gene_abundance_factor_mu}{numeric}

\item{gene_abundance_factor_var}{numeric}

\item{gene_hit_rate}{numeric}

\item{n_wells_per_screen}{integer}

\item{well_abundance_factor_mu}{numeric}

\item{well_abundance_factor_var}{numeric}

\item{n_genes_per_well}{integer}

\item{n_cells_per_well_mu}{numeric}

\item{n_cells_per_well_var}{numeric}

\item{class_pos_mu}{numeric}

\item{class_pos_var}{numeric}

\item{class_neg_mu}{numeric}

\item{class_neg_var}{numeric}

\item{verbose}{= FALSE}
}
\description{
For each combination of given parameters, generate a simulated dataset
fit a model to it and estimate the predictive accuracy.
}
\examples{
\dontrun{
  scores <- spaCRPower::scan_parameters(
    n_genes_per_library = 200,
    gene_abundance_factor_mu = 1,
    gene_abundance_factor_var = .01,
    gene_hit_rate = 0.05,
    n_wells_per_screen = seq(30, 200, length.out = 35),
    well_abundance_factor_mu = 1,
    well_abundance_factor_var = .01,
    n_genes_per_well = c(1, 3, 10, 30),
    n_cells_per_well_mu = 850,
    n_cells_per_well_var = 1000,
    class_pos_mu = 0.99,
    class_pos_var = 0.0001,
    class_neg_mu = 0.01,
    class_neg_var = 0.0001,
    progress_file = "parameter_scan.tsv",
    verbose = TRUE,
    refresh = 0)
}

}
